Add mpasa60 to namelist defaults and definition files
This commit adds mpasa60 surfdata, SCRIP file, and namelist definition and
defaults to CTSM.

It should be noted, that as of this commit, the mpasa60 surfdata was created
using the tx0.1v2 mask and not the gx1v7 mask as the other MPAS grids have
done. This may be incorrect, and may need to be changed.

Per instructions from Erik Kluzek, this data was created with the
ctsm5.1.dev038 tag.